feat: CheckLocation add
