+++ src/App.tsx
@@ -7,12 +7,14 @@ import { RecoilRoot } from 'recoil';
 import { Layout } from './components/Layout';
 
 import '@karrotframe/navigator/index.css';
+import CheckLocation from './components/common/CheckLocation';
 
 function App() {
   return (
     <ThemeProvider theme={theme}>
       <SWRConfig>
         <RecoilRoot>
+          <CheckLocation />
           <div
             style={{
               display: 'flex',
