+++ src/components/Layout/index.tsx
@@ -19,28 +19,23 @@ export const Layout = () => {
   const { os } = parser(window.navigator.userAgent);
   const isCupertino = os.name === IOS;
   const [navHandler, setNavHandler] = useRecoilState(navState);
-  console.log('nav' + navHandler.NAVIGATION);
-  const hideNavigation = () => {
-    if (location.pathname.includes('/recruitment/detail')) {
-      setNavHandler({ ...navHandler, NAVIGATION: false });
-    } else {
-      setNavHandler({ ...navHandler, NAVIGATION: true });
-    }
-  };
-  useEffect(() => {
-    hideNavigation();
-  }, []);
+
   return (
-    <Navigator
-      className={navHandler.NAVIGATION ? '' : 'navigateClean'}
-      theme={isCupertino ? 'Cupertino' : 'Android'}
-      onClose={() => console.log('onClose')}
-    >
-      <Screen path={'/list/all'} component={All} />
-      <Screen path={'/list/recommend'} component={Recommend} />
-      <Screen path={'/list/book-marked'} component={Bookmarked} />
-      <Screen path={'/recruitment/detail/:id'} component={RecruitmentDetail} />
-      <Screen path={'/resume/:id'} component={RecruitmentDetail} />
-    </Navigator>
+    <>
+      <Navigator
+        className={navHandler.NAVIGATION ? '' : 'navigateClean'}
+        theme={isCupertino ? 'Cupertino' : 'Android'}
+        onClose={() => console.log('onClose')}
+      >
+        <Screen path={'/list/all'} component={All} />
+        <Screen path={'/list/recommend'} component={Recommend} />
+        <Screen path={'/list/book-marked'} component={Bookmarked} />
+        <Screen
+          path={'/recruitment/detail/:id'}
+          component={RecruitmentDetail}
+        />
+        <Screen path={'/resume/:id'} component={RecruitmentDetail} />
+      </Navigator>
+    </>
   );
 };
