+++ src/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import { BrowserRouter } from 'react-router-dom';
+import CheckLocation from './components/common/CheckLocation';
 
 ReactDOM.render(
   <React.StrictMode>
